requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
fastapi-cors
pydantic<2.0.0
python-multipart
pydantic[email]
bcrypt==3.2.0
boto3
python-dotenv
passlib[bcrypt]
fastapi-login
python-jose
passlib
starlette
itsdangerous==2.2.0
nltk
requests